chore: move action struct to ha folder
