+++ src/ha/panels/lovelace/editor/structs/action-struct.ts
@@ -1,4 +1,16 @@
-import { object, string, union, boolean, optional, array, literal, enums, type } from "superstruct";
+import {
+    array,
+    boolean,
+    dynamic,
+    enums,
+    literal,
+    object,
+    optional,
+    string,
+    type,
+    union,
+} from "superstruct";
+import { BaseActionConfig } from "../../../../data/lovelace";
 
 const actionConfigStructUser = object({
     user: string(),
@@ -22,6 +34,7 @@ const actionConfigStructService = object({
     action: literal("call-service"),
     service: string(),
     service_data: optional(object()),
+    data: optional(object()),
     target: optional(
         object({
             entity_id: optional(union([string(), array(string())])),
@@ -47,10 +60,23 @@ export const actionConfigStructType = object({
     confirmation: optional(actionConfigStructConfirmation),
 });
 
-export const actionConfigStruct = union([
-    actionConfigStructType,
-    actionConfigStructUrl,
-    actionConfigStructNavigate,
-    actionConfigStructService,
-    actionConfigStructCustom,
-]);
+export const actionConfigStruct = dynamic<any>((value) => {
+    if (value && typeof value === "object" && "action" in value) {
+        switch ((value as BaseActionConfig).action!) {
+            case "call-service": {
+                return actionConfigStructService;
+            }
+            case "fire-dom-event": {
+                return actionConfigStructCustom;
+            }
+            case "navigate": {
+                return actionConfigStructNavigate;
+            }
+            case "url": {
+                return actionConfigStructUrl;
+            }
+        }
+    }
+
+    return actionConfigStructType;
+});
