+++ src/shared/config/actions-config.ts
@@ -1,6 +1,6 @@
-import { ActionConfig } from "custom-card-helpers";
 import { object, optional } from "superstruct";
-import { actionConfigStruct } from "../../utils/action-struct";
+import { ActionConfig } from "../../ha/data/lovelace";
+import { actionConfigStruct } from "../../ha/panels/lovelace/editor/structs/action-struct";
 
 export const actionsSharedConfigStruct = object({
     tap_action: optional(actionConfigStruct),
@@ -8,7 +8,6 @@ export const actionsSharedConfigStruct = object({
     double_tap_action: optional(actionConfigStruct),
 });
 
-// TODO : fix action config type and struct
 export type ActionsSharedConfig = {
     tap_action?: ActionConfig;
     hold_action?: ActionConfig;
